cache: split code into db & gh submodules
